+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Add tests for raster module (@kysolvik, #32)
 * Added tests for base module -- legend check methods (@kysolvik, #38)
 * Add conftest.py to centralize pytest fixtures (@kysolvik, #35)
 * Fix issue with pip 19.1 breaking appveyor build (@kysolvik, #46)